+++ sdk/src/utils/test.rs
@@ -49,7 +49,7 @@ pub(crate) fn create_mock_file_with_name_path(name: &str, path: &str) -> File {
 }
 
 #[allow(dead_code)]
-pub(crate) fn create_mock_function(id: u32) -> Function {
+pub(crate) fn create_mock_function(id: u128) -> Function {
     Function {
         id,
         location: create_mock_location(),
@@ -63,7 +63,7 @@ pub(crate) fn create_mock_function(id: u32) -> Function {
 
 #[allow(dead_code)]
 pub(crate) fn create_mock_function_with_parameters(
-    id: u32,
+    id: u128,
     parameters: Vec<Rc<FnParameter>>,
 ) -> Function {
     Function {
@@ -78,7 +78,7 @@ pub(crate) fn create_mock_function_with_parameters(
 }
 
 #[allow(dead_code)]
-pub(crate) fn create_mock_contract(id: u32) -> Contract {
+pub(crate) fn create_mock_contract(id: u128) -> Contract {
     Contract {
         id,
         name: "TestContract".to_string(),
@@ -89,7 +89,7 @@ pub(crate) fn create_mock_contract(id: u32) -> Contract {
 
 #[allow(dead_code)]
 pub(crate) fn create_mock_contract_with_inner_struct(
-    id: u32,
+    id: u128,
     name: String,
     location: Location,
 ) -> Contract {
